Add files via upload
A text editor is a computer program that enables users to create, change, or edit plain text files (i.e., files with the suffix . txt). They're often used to craft complex code for websites; read, create and edit source code; or build text files. I had made this project with the help of Tkinter only. Tkinter is the inbuilt python module that is used to create GUI applications.

Tkinter with some basic concept of Python Programming is sufficient to make this project.